requirements.txt
fastapi[all]
pymongo
passlib
bcrypt
python-jose
fastapi-sessions
